move back op-stack bridges to original locations
