+++ src/clients/bridges/op-stack/optimism/SnxOptimismBridge.ts
@@ -1,20 +1,20 @@
-import { Contract, BigNumber, paginatedEventQuery, EventSearchConfig, Signer, Provider } from "../../../utils";
-import { CONTRACT_ADDRESSES } from "../../../common";
-import { OpStackBridge, BridgeTransactionDetails, OpStackEvents } from "./OpStackBridgeInterface";
+import { Contract, BigNumber, paginatedEventQuery, EventSearchConfig, Signer, Provider } from "../../../../utils";
+import { CONTRACT_ADDRESSES } from "../../../../common";
+import { OpStackBridge, BridgeTransactionDetails, OpStackEvents } from "../OpStackBridgeInterface";
 
-export class DaiOptimismBridge extends OpStackBridge {
+export class SnxOptimismBridge extends OpStackBridge {
   private readonly l1Bridge: Contract;
   private readonly l2Bridge: Contract;
 
   constructor(l2chainId: number, hubChainId: number, l1Signer: Signer, l2SignerOrProvider: Signer | Provider) {
     super(l2chainId, hubChainId, l1Signer, l2SignerOrProvider, [
-      CONTRACT_ADDRESSES[hubChainId].daiOptimismBridge.address,
+      CONTRACT_ADDRESSES[hubChainId].snxOptimismBridge.address,
     ]);
 
-    const { address: l1Address, abi: l1Abi } = CONTRACT_ADDRESSES[hubChainId].daiOptimismBridge;
+    const { address: l1Address, abi: l1Abi } = CONTRACT_ADDRESSES[hubChainId].snxOptimismBridge;
     this.l1Bridge = new Contract(l1Address, l1Abi, l1Signer);
 
-    const { address: l2Address, abi: l2Abi } = CONTRACT_ADDRESSES[l2chainId].daiOptimismBridge;
+    const { address: l2Address, abi: l2Abi } = CONTRACT_ADDRESSES[l2chainId].snxOptimismBridge;
     this.l2Bridge = new Contract(l2Address, l2Abi, l2SignerOrProvider);
   }
 
@@ -23,24 +23,27 @@ export class DaiOptimismBridge extends OpStackBridge {
     l1Token: string,
     l2Token: string,
     amount: BigNumber,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     l2Gas: number
   ): BridgeTransactionDetails {
     return {
       contract: this.l1Bridge,
-      method: "depositERC20",
-      args: [l1Token, l2Token, amount, l2Gas, "0x"],
+      method: "depositTo",
+      args: [toAddress, amount],
     };
   }
 
   async queryL1BridgeInitiationEvents(
     l1Token: string,
-    fromAddress: string,
+    toAddress: string,
     eventConfig: EventSearchConfig
   ): Promise<OpStackEvents> {
+    // @dev For the SnxBridge, only the `toAddress` is indexed on the L2 event so we treat the `fromAddress` as the
+    // toAddress when fetching the L1 event.
     return {
       [this.resolveL2TokenAddress(l1Token)]: await paginatedEventQuery(
         this.l1Bridge,
-        this.l1Bridge.filters.ERC20DepositInitiated(l1Token, undefined, fromAddress),
+        this.l1Bridge.filters.DepositInitiated(undefined, toAddress),
         eventConfig
       ),
     };
@@ -48,13 +51,13 @@ export class DaiOptimismBridge extends OpStackBridge {
 
   async queryL2BridgeFinalizationEvents(
     l1Token: string,
-    fromAddress: string,
+    toAddress: string,
     eventConfig: EventSearchConfig
   ): Promise<OpStackEvents> {
     return {
       [this.resolveL2TokenAddress(l1Token)]: await paginatedEventQuery(
         this.l2Bridge,
-        this.l2Bridge.filters.DepositFinalized(l1Token, undefined, fromAddress),
+        this.l2Bridge.filters.DepositFinalized(toAddress),
         eventConfig
       ),
     };
